% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcurate-package.R
\docType{package}
\name{ttcurate-package}
\alias{ttcurate}
\alias{ttcurate-package}
\title{ttcurate: What the Package Does}
\description{
Helper functions and guidance for curating datasets for the 'TidyTuesday' weekly data project. 'TidyTuesday' is a project by the 'Data Science Learning Community' in which they post a weekly dataset in a public data repository (\url{https://github.com/rfordatascience/tidytuesday}) for people to analyze and visualize. This package includes tools to explore past submissions, analyze dataset sizes, and follow best practices for data preparation and documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dslc-io/ttcurate}
  \item \url{https://dslc-io.github.io/ttcurate/}
  \item Report bugs at \url{https://github.com/dslc-io/ttcurate/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
